improvements from code review
