+++ src/adapter/mod.rs
@@ -21,7 +21,7 @@ mod enum_variant;
 mod optimizations;
 mod origin;
 mod properties;
-mod typename;
+mod rust_type_name;
 mod vertex;
 
 #[cfg(test)]
